# docs/aliyun_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "阿里云接口文档",
        "contact": {},
        "version": "1.0 版本"
    },
    "basePath": "/app/aliyun/api/",
    "paths": {
        "/config/aliyun": {
            "get": {
                "description": "查询阿里云配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "查询阿里云配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-internal_Options"
                        }
                    }
                }
            },
            "post": {
                "description": "修改阿里云配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "修改阿里云配置",
                "parameters": [
                    {
                        "description": "阿里云配置",
                        "name": "cfg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.Options"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int"
                        }
                    }
                }
            }
        },
        "/sms/count": {
            "post": {
                "description": "查询短消息数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "查询短消息数量",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/sms/create": {
            "post": {
                "description": "创建短消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "创建短消息",
                "parameters": [
                    {
                        "description": "短消息信息",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Sms"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Sms"
                        }
                    }
                }
            }
        },
        "/sms/list": {
            "get": {
                "description": "查询短消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "查询短消息",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Sms"
                        }
                    }
                }
            }
        },
        "/sms/search": {
            "post": {
                "description": "查询短消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "查询短消息",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Sms"
                        }
                    }
                }
            }
        },
        "/sms/{id}": {
            "get": {
                "description": "获取短消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "获取短消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "短消息ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Sms"
                        }
                    }
                }
            }
        },
        "/sms/{id}/delete": {
            "get": {
                "description": "删除短消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "删除短消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "短消息ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Sms"
                        }
                    }
                }
            }
        },
        "/subscriber/count": {
            "post": {
                "description": "查询订阅数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriber"
                ],
                "summary": "查询订阅数量",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/subscriber/create": {
            "post": {
                "description": "创建订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriber"
                ],
                "summary": "创建订阅",
                "parameters": [
                    {
                        "description": "订阅信息",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Subscriber"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Subscriber"
                        }
                    }
                }
            }
        },
        "/subscriber/export": {
            "get": {
                "description": "导出订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "product"
                ],
                "summary": "导出订阅",
                "responses": {}
            }
        },
        "/subscriber/import": {
            "post": {
                "description": "导入订阅",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "导入订阅",
                "parameters": [
                    {
                        "type": "file",
                        "description": "压缩包",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/subscriber/list": {
            "get": {
                "description": "查询订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriber"
                ],
                "summary": "查询订阅",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Subscriber"
                        }
                    }
                }
            }
        },
        "/subscriber/search": {
            "post": {
                "description": "查询订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriber"
                ],
                "summary": "查询订阅",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Subscriber"
                        }
                    }
                }
            }
        },
        "/subscriber/{id}": {
            "get": {
                "description": "获取订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriber"
                ],
                "summary": "获取订阅",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订阅ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Subscriber"
                        }
                    }
                }
            },
            "post": {
                "description": "修改订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriber"
                ],
                "summary": "修改订阅",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订阅ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "订阅信息",
                        "name": "subscriber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Subscriber"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Subscriber"
                        }
                    }
                }
            }
        },
        "/subscriber/{id}/delete": {
            "get": {
                "description": "删除订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriber"
                ],
                "summary": "删除订阅",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订阅ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Subscriber"
                        }
                    }
                }
            }
        },
        "/subscriber/{id}/disable": {
            "get": {
                "description": "禁用订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriber"
                ],
                "summary": "禁用订阅",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订阅ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Subscriber"
                        }
                    }
                }
            }
        },
        "/subscriber/{id}/enable": {
            "get": {
                "description": "启用订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriber"
                ],
                "summary": "启用订阅",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订阅ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Subscriber"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ParamSearch": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": true
                },
                "keyword": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "skip": {
                    "type": "integer"
                },
                "sort": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.ReplyData-int": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyData-int64": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyData-internal_Options": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal.Options"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyData-types_Sms": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Sms"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyData-types_Subscriber": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Subscriber"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyList-types_Sms": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Sms"
                    }
                },
                "error": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.ReplyList-types_Subscriber": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Subscriber"
                    }
                },
                "error": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal.Options": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "sign": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                }
            }
        },
        "types.Sms": {
            "type": "object",
            "properties": {
                "cellphone": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "result": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "types.Subscriber": {
            "type": "object",
            "properties": {
                "cellphone": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}